requirements.txt
paho-mqtt
loguru
pytest
python-dotenv
requests